[random fix]: i really dont know whats wrong with the code

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,5 +1,4 @@
 #include "function_pointers.h"
-#include <stdlib.h>
 
 /**
  * array_iterator - executes a function givenn as a parameter on each element
@@ -13,7 +12,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i;
+	unsigned int i;
 
 	if (array == NULL || action == NULL)
 		return;
